Add another explicit cast for Qt 5.9 (and presumably earlier)

+++ test/unit/lib/testabstractpokitservice.cpp
@@ -46,7 +46,7 @@ void TestAbstractPokitService::pokitProduct()
 {
     QFETCH(PokitProduct, product);
     MockPokitService service(nullptr);
-    QCOMPARE(static_cast<quint8>(service.pokitProduct()), 255u); // ie invalid / unset by default.
+    QCOMPARE(static_cast<quint8>(service.pokitProduct()), (quint8)255u); // ie invalid / unset by default.
     service.setPokitProduct(product);
     QVERIFY(service.pokitProduct() == product);
 }
